scripts/requirements.txt
supabase==2.4.4
python-dotenv==1.0.0
